+++ src/spyglass/utils/mixins/analysis.py
@@ -380,10 +380,41 @@ class AnalysisMixin(BaseMixin):
         path : str
             The path for the analysis NWB file.
         """
-        abs_path = cls.__get_file_parent(fname) / fname
-        return (
-            abs_path.relative_to(cls()._analysis_dir) if relative else abs_path
-        )
+        analysis_dir = cls()._analysis_dir
+        old_format = Path(analysis_dir) / fname  # Flat stored, see #1565
+
+        if old_format.exists():
+            abs_path = old_format
+        else:
+            abs_path = cls.__get_file_parent(fname) / fname
+
+        return abs_path.relative_to(analysis_dir) if relative else abs_path
+
+    @classmethod
+    def _get_analysis_file_paths(
+        cls, fnames: list, relative: bool = False, as_str: bool = True
+    ) -> list:
+        """Get the paths for a list of analysis NWB files.
+
+        Parameters
+        ----------
+        fnames : list
+            A list of analysis NWB file names.
+        relative : bool, Optional
+            If true, return paths relative to analysis_dir. Defaults False.
+        as_str : bool, Optional
+            If true, return paths as strings. If false, return as Path objects.
+
+        Returns
+        -------
+        paths : list
+            A list of paths for the specified analysis NWB files.
+        """
+        ret = [
+            cls.__get_analysis_path(fname, relative=relative)
+            for fname in fnames
+        ]
+        return [str(path) for path in ret] if as_str else ret
 
     @classmethod
     def copy(cls, nwb_file_name: str):